Move test to a new home
